tied up lose ends
